+++ src/app/jobList/jobDetail/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const JobDetail = () => {
+  return (
+    <div>구인페이지 상세</div>
+  )
+}
+
+export default JobDetail
