+++ src/app/jobList/new/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const JobListNew = () => {
+  return (
+    <div>구인 등록페이지</div>
+  )
+}
+
+export default JobListNew
