+++ src/app/jobList/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const JobList = () => {
+  return (
+    <div>구인리스트</div>
+  )
+}
+
+export default JobList
