+++ src/app/page.tsx
@@ -1,5 +1,8 @@
 // 메인페이지
-import MainPage from "./_components/MainPage";
+
+import MainPage from "@/components/main/MainPage";
+
+
 
 export default function Home() {
   return (
